Article shorts: admin Generate-short button + TS queue (phase 4)
- short-render-queue.ts: TS mirror of the short_renders queue (enqueue idempotent
  on story_id + narration/length hash, claim, get, latest, finish/fail, progress)
- shorts-options.ts: narration vibe + length preset lists for the picker
- videos/[id]/actions.ts: queueShortRender (admin-only, daily-capped) +
  getShortRenderStatusAction + latestShortRenderAction
- ShortRenderControl.tsx: vibe + length pickers, Generate-short button, progress
  bar (polls status), inline player when done
- EditorClient.tsx: mount the control as a band under the video editor header
- schema.ts: short_renders.progress INTEGER to match VIDEO_RENDERS / ColType

Local flow now works end to end: button enqueues, pipeline/short_render_worker.py
drains and renders. Phase 3 (Vercel cron + Cloud Run) still to wire for prod.

+++ lorewire-app/src/lib/schema.ts
@@ -386,7 +386,10 @@ export const SHORT_RENDERS: Table = {
     { name: "length_preset", type: "TEXT" },
     { name: "status", type: "TEXT" },
     { name: "phase", type: "TEXT" },
-    { name: "progress", type: "REAL" },
+    // progress matches VIDEO_RENDERS (the ColType union has no REAL); the Python
+    // store.py holds the 0..1 value as REAL and owns the prod schema, SQLite is
+    // dynamically typed, so the TS column type here is only the local-dev mirror.
+    { name: "progress", type: "INTEGER" },
     { name: "error", type: "TEXT" },
     { name: "output_url", type: "TEXT" },
     { name: "requested_by", type: "TEXT" },
